requirement.txt
1.numpy
2.pandas
3.seaborn
4.matplotlib.pyplot
5.base64
6.streamlit
7.joblin
8.from sklearn.linear_model import LogisticRegression
9.from sklearn.tree import DecisionTreeRegressor
,
GradientBoostingRegressor, AdaBoostRegressor
10.from sklearn.ensemble import RandomForestClassifier
11.from xgboost import XGBClassifier